chore(fqdn): fqdn move to specific folder and update documentation
