+++ charts/external-dns/tests/deployment-flags_test.yaml
@@ -164,4 +164,3 @@ tests:
     asserts:
      - failedTemplate:
           errorMessage: "'txtPrefix' and 'txtSuffix' are mutually exclusive"
-
