+++ source/openshift_route.go
@@ -34,6 +34,7 @@ import (
 	"k8s.io/client-go/tools/cache"
 
 	"sigs.k8s.io/external-dns/endpoint"
+	"sigs.k8s.io/external-dns/source/fqdn"
 )
 
 // ocpRouteSource is an implementation of Source for OpenShift Route objects.
@@ -65,7 +66,7 @@ func NewOcpRouteSource(
 	labelSelector labels.Selector,
 	ocpRouterName string,
 ) (Source, error) {
-	tmpl, err := parseTemplate(fqdnTemplate)
+	tmpl, err := fqdn.ParseTemplate(fqdnTemplate)
 	if err != nil {
 		return nil, err
 	}
